Add Developer Menu option for resetting Firebase Flags
