+++ podcasts/DeveloperMenu.swift
@@ -36,6 +36,12 @@ struct DeveloperMenu: View {
                         PodcastManager.shared.unsubscribe(podcast: podcast)
                     }
                 }
+
+                Button("Force Reload Feature Flags") {
+                    FirebaseManager.refreshRemoteConfig(expirationDuration: 0) { _ in
+                        (UIApplication.shared.delegate as? AppDelegate)?.updateRemoteFeatureFlags()
+                    }
+                }
             }
 
             Section {
